Improvement of Binary(.cpp/.hpp) to handle special case of Spotless loading itself. Compile with gcc 11. Added 'Stop' action.

+++ SimpleDebug/Strings.hpp
@@ -78,7 +78,7 @@ public:
     }
     int getInt() { //this is a little tricky because of problems with c++ and adtools
         bool negative = false;
-        unsigned int result = 0;
+        unsigned long long result = 0;
         char c = ss.peek();
         if(c == '-') {
             negative = true;
